lbft27j

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: batuhankumcu <batuhankumcu@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/07 16:39:02 by batuhankumc       #+#    #+#             */
-/*   Updated: 2026/01/23 14:53:38 by batuhankumc      ###   ########.fr       */
+/*   Updated: 2026/01/25 15:29:49 by batuhankumc      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
+	size_t	i;
 	int	a;
 
 	i = 0;
